refactor: Move remaining VM function tests to separate files
Extract inline test modules from math_functions.rs, map_functions.rs,
number_functions.rs, and boolean_functions.rs into dedicated test files
under src/vm/tests/. Maintains all 71 tests (35 math + 14 map + 14 number
+ 8 boolean) with identical behavior.
