+++ src/vm/tests/mod.rs
@@ -5,3 +5,7 @@ mod string_functions;
 mod set_functions;
 mod array_functions;
 mod file_functions;
+mod math_functions;
+mod map_functions;
+mod number_functions;
+mod boolean_functions;
